Added exceptions and local dictionary for roles db

+++ data.py
@@ -1,6 +1,6 @@
 from tinydb import TinyDB, Query
 import warnings
-from exceptions import InvalidInsertRole, DuplicateRole
+from exceptions import InvalidInsertRole, InvalidRemoveRole, DuplicateRole
 
 standard_roles = [
     {"category": "Character roles", "role": "Alisa Main"},
@@ -21,21 +21,32 @@ standard_roles = [
     {"category": "Misc roles", "role": "Streamer"}
 ]
 
-class DataBase:
+class RolesDataBase:
+    roles_dict = {}
     def __init__(self):
         self.roles_db = TinyDB("data/roles.json")
         self.query = Query()
         if len(self.roles_db) == 0:
             warnings.warn("data/roles.json file seems to be empty, filling it up pre-configured standard roles.")
             self.roles_db.insert_multiple(standard_roles)
-    
+        self.db_to_dict()
+
+    def db_to_dict(self):
+        self.roles_dict.clear()
+        db = self.get_database()
+        for entry in db:
+            self.roles_dict[entry['category']] = []
+        for entry in db:
+            self.roles_dict[entry['category']].append(entry['role'])
     def try_out(self):
-        self.roles_db.truncate()
-        self.roles_db.insert_multiple(standard_roles)
+        # self.roles_db.truncate()
+        # self.roles_db.insert_multiple(standard_roles)
         #self.insert_role("Character roles", "Alisa Main")
+        # print(self.roles_dict)
+        return
     def is_duplicate(self, role):
         return self.roles_db.contains(self.query['role']==role)
-    def insert_role(self, category, role):
+    def add_role(self, category, role):
         if(role == 'Admin' or role == 'Mod'):
             raise InvalidInsertRole(f"I'm not allowed to assign {role} to regular members.")
         elif(not(self.is_duplicate(role))):
@@ -43,19 +54,21 @@ class DataBase:
         else:
             raise DuplicateRole(f"The role {role} already exists in the list.")
     def remove_role(self, role):
+        if(not(self.is_duplicate(role))):
+            raise InvalidRemoveRole(f"The role {role} is not present in the database.")
         self.roles_db.remove(self.query.role == role)
     def get_database(self):
         return self.roles_db.all()
-    def get_roles(self):
-        table = self.roles_db.table('role')
-        roles = table.all()
-        role_values =[role_value['role'] for role in roles]
+    def get_dictionary(self):
+        return self.roles_dict
+    #TODO Perhaps change to outputting with dictionary results
     def search_category(self, category_val):
         result = self.roles_db.search(self.query.category == category_val)
         return result
 
-test_db = DataBase()
+test_db = RolesDataBase()
 
 test_db.try_out()
-test_result = test_db.get_roles()
-print(test_result)
+# test_result = test_db.get_roles()
+# print(test_result)
+print(test_db.search_category('Character roles'))
